Made Lots of Changes
Got the basics of a file upload system working,

Updated the look of buttons,

Got a search bar working for clients and suppliers

Got the submission to create new suppliers and clients communicating with the database.

There is still a lot of work to do.  All of these additions work when data is entered perfectly.  If something is off however the server will crash.  Need error correction.